--- a/src/libs/E2E/tests/appStartTimeTest.e2e.ts
+++ b/src/libs/E2E/tests/appStartTimeTest.e2e.ts
@@ -26,7 +26,8 @@ const test = () => {
                 E2EClient.submitTestResults({
                     branch: Config.E2E_BRANCH,
                     name: `App start ${metric.name}`,
-                    duration: metric.duration,
+                    metric: metric.duration,
+                    unit: 'ms',
                 }),
             ),
         )