--- a/src/libs/E2E/tests/linkingTest.e2e.ts
+++ b/src/libs/E2E/tests/linkingTest.e2e.ts
@@ -75,7 +75,8 @@ const test = (config: NativeConfig) => {
                 E2EClient.submitTestResults({
                     branch: Config.E2E_BRANCH,
                     name: 'Comment linking',
-                    duration: entry.duration,
+                    metric: entry.duration,
+                    unit: 'ms',
                 });
 
                 switchReportResolve();