--- a/src/libs/E2E/types.ts
+++ b/src/libs/E2E/types.ts
@@ -33,6 +33,8 @@ type TestModule = {default: Test};
 
 type Tests = Record<ValueOf<typeof E2EConfig.TEST_NAMES>, Test>;
 
+type Unit = 'ms' | 'MB' | '%' | 'renders' | 'FPS';
+
 type TestResult = {
     /** Name of the test */
     name: string;
@@ -40,8 +42,8 @@ type TestResult = {
     /** The branch where test were running */
     branch?: string;
 
-    /** Duration in milliseconds */
-    duration?: number;
+    /** The numeric value of the measurement */
+    metric?: number;
 
     /** Optional, if set indicates that the test run failed and has no valid results. */
     error?: string;
@@ -52,8 +54,8 @@ type TestResult = {
      */
     isCritical?: boolean;
 
-    /** Render count */
-    renderCount?: number;
+    /** The unit of the measurement */
+    unit?: Unit;
 };
 
-export type {SigninParams, IsE2ETestSession, NetworkCacheMap, NetworkCacheEntry, TestConfig, TestResult, TestModule, Tests};
+export type {SigninParams, IsE2ETestSession, NetworkCacheMap, NetworkCacheEntry, TestConfig, TestResult, TestModule, Tests, Unit};