--- a/tests/e2e/ADDING_TESTS.md
+++ b/tests/e2e/ADDING_TESTS.md
@@ -72,7 +72,8 @@ const test = () => {
       // ... do something with the measurements
       E2EClient.submitTestResults({
           name: "Navigate to report",
-          duration: measurement.duration,
+          metric: measurement.duration,
+          unit: 'ms',
       }).then(E2EClient.submitTestDone)
   });