--- a/tests/e2e/README.md
+++ b/tests/e2e/README.md
@@ -148,7 +148,8 @@ const someDurationWeCollected = // ...
 
 E2EClient.submitTestResults({
     name: 'My test name',
-    duration: someDurationWeCollected,
+    metric: someDurationWeCollected,
+    unit: 'ms',
 });
 ```