<!-- If necessary, assign reviewers that know the area or changes well. Feel free to tag any additional reviewers you see fit. -->

### Details

Follow up PR for https://github.com/Expensify/App/pull/43482 In this PR I'm adding different units for different types of measurement.

> [!CAUTION]
> Highly desirable to merge this PR straight before a new release, since it contains pretty a lot of breaking changes and old `.apk`s will be incompatible with new test runner. Should be merged after https://github.com/Expensify/App/pull/43482

|Before|After|
|-------|-----|
|<img width="888" alt="image" src="https://github.com/Expensify/App/assets/22820318/1308cad3-5c5d-445f-9ee2-95f6d4c6dea2">|<img width="837" alt="image" src="https://github.com/Expensify/App/assets/22820318/f6f5133a-25c6-4ab3-97d7-97e72ac8eb40">|

Below you can find an example of generated markdown:

## Performance Comparison Report 📊

### Significant Changes To Duration
_There are no entries_


### Meaningless Changes To Duration
<details>
<summary>Show entries</summary>

| Name                                          | Duration    |
| --------------------------------------------- | ----------- |
| App start nativeLaunchEnd_To_appCreationStart | 83.375 ms   |
| App start nativeLaunch                        | 89.375 ms   |
| App start appCreation                         | 90.125 ms   |
| App start appCreationEnd_To_contentAppeared   | 647.875 ms  |
| App start contentAppeared_To_screenTTI        | 1267.619 ms |
| App start runJsBundle                         | 477.125 ms  |
| App start TTI                                 | 2178.369 ms |
| App start regularAppStart                     | 0.025 ms    |
| App start time (CPU)                          | 156.901 %   |
| App start time (FPS)                          | 59.784 FPS  |
| App start time (RAM)                          | 356.692 MB  |
</details>


<details>
<summary>Show details</summary>

| Name                                          | Duration                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        |
| --------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
| App start nativeLaunchEnd_To_appCreationStart | **Baseline**<br/>Mean: 100.750 ms<br/>Stdev: 56.740 ms (56.3%)<br/>Runs: 59 67 72 78 85 90 109 246<br/><br/>**Current**<br/>Mean: 83.375 ms<br/>Stdev: 11.916 ms (14.3%)<br/>Runs: 60 79 80 81 83 84 100 100                                                                                                                                                                                                                                                                                    |
| App start nativeLaunch                        | **Baseline**<br/>Mean: 85.375 ms<br/>Stdev: 4.794 ms (5.6%)<br/>Runs: 78 80 83 85 86 88 89 94<br/><br/>**Current**<br/>Mean: 89.375 ms<br/>Stdev: 10.012 ms (11.2%)<br/>Runs: 77 79 87 88 89 89 94 112                                                                                                                                                                                                                                                                                          |
| App start appCreation                         | **Baseline**<br/>Mean: 97.875 ms<br/>Stdev: 16.922 ms (17.3%)<br/>Runs: 67 90 90 94 96 108 109 129<br/><br/>**Current**<br/>Mean: 90.125 ms<br/>Stdev: 12.634 ms (14.0%)<br/>Runs: 74 83 84 85 91 91 93 120                                                                                                                                                                                                                                                                                     |
| App start appCreationEnd_To_contentAppeared   | **Baseline**<br/>Mean: 622.125 ms<br/>Stdev: 49.798 ms (8.0%)<br/>Runs: 562 577 590 592 635 636 661 724<br/><br/>**Current**<br/>Mean: 647.875 ms<br/>Stdev: 61.310 ms (9.5%)<br/>Runs: 568 592 601 606 657 694 726 739                                                                                                                                                                                                                                                                         |
| App start contentAppeared_To_screenTTI        | **Baseline**<br/>Mean: 1305.782 ms<br/>Stdev: 133.131 ms (10.2%)<br/>Runs: 1156.5970499999821 1187.303460999974 1206.6081209999975 1241.2006399999955 1265.4872229999746 1362.3920100000105 1497.405487000011 1529.264254000038<br/><br/>**Current**<br/>Mean: 1267.619 ms<br/>Stdev: 105.689 ms (8.3%)<br/>Runs: 1140.6160490000038 1201.7801089999848 1205.4019809999736 1207.8677119999775 1233.2525310000055 1293.425217000011 1368.7890729999635 1489.8198379999958                        |
| App start runJsBundle                         | **Baseline**<br/>Mean: 460.250 ms<br/>Stdev: 32.771 ms (7.1%)<br/>Runs: 430 439 441 442 452 454 489 535<br/><br/>**Current**<br/>Mean: 477.125 ms<br/>Stdev: 55.379 ms (11.6%)<br/>Runs: 408 422 422 462 474 528 544 557                                                                                                                                                                                                                                                                        |
| App start TTI                                 | **Baseline**<br/>Mean: 2211.907 ms<br/>Stdev: 211.166 ms (9.5%)<br/>Runs: 2002.5970499999821 2092.4872229999746 2093.2006399999955 2119.303460999974 2126.6081209999975 2161.3920100000105 2412.405487000011 2687.264254000038<br/><br/>**Current**<br/>Mean: 2178.369 ms<br/>Stdev: 167.123 ms (7.7%)<br/>Runs: 2005.6160490000038 2038.8677119999775 2039.7801089999848 2054.4019809999736 2153.2525310000055 2301.425217000011 2345.7890729999635 2487.819837999996                          |
| App start regularAppStart                     | **Baseline**<br/>Mean: 0.025 ms<br/>Stdev: 0.005 ms (19.3%)<br/>Runs: 0.020548000000417233 0.021524999989196658 0.022136000043246895 0.023803999996744096 0.023803999996744096 0.024209999945014715 0.026285999978426844 0.03682500007562339<br/><br/>**Current**<br/>Mean: 0.025 ms<br/>Stdev: 0.004 ms (16.9%)<br/>Runs: 0.02172800002153963 0.022054000000935048 0.02229799999622628 0.023111999966204166 0.023763000033795834 0.024169999989680946 0.026082000054884702 0.03535999997984618 |
| App start time (CPU)                          | **Baseline**<br/>Mean: 154.588 %<br/>Stdev: 3.032 % (2.0%)<br/>Runs: 150.17349546001842 151.75041261148448 153.19055133810605 157.0538571261075 157.18023931776335 158.1773071731539<br/><br/>**Current**<br/>Mean: 156.901 %<br/>Stdev: 4.600 % (2.9%)<br/>Runs: 148.5396991273787 152.86475272323335 154.59480973324114 155.10880662339412 159.736714517591 160.4016648363146 160.82644393938259 163.13438238127955                                                                           |
| App start time (FPS)                          | **Baseline**<br/>Mean: 57.844 FPS<br/>Stdev: 1.703 FPS (2.9%)<br/>Runs: 55.107489304472004 56.0982928508171 57.39432223546879 57.52587874695508 59.224918773637285 59.58690738429183 59.96670772248177<br/><br/>**Current**<br/>Mean: 59.784 FPS<br/>Stdev: 0.452 FPS (0.8%)<br/>Runs: 58.77512689419158 59.92911343561765 60 60 60 60                                                                                                                                                          |
| App start time (RAM)                          | **Baseline**<br/>Mean: 352.439 MB<br/>Stdev: 8.207 MB (2.3%)<br/>Runs: 340.0572916666667 340.7174479166667 349.75390625 351.2109375 353.88125 360.4697265625 361.0322265625 362.3916015625<br/><br/>**Current**<br/>Mean: 356.692 MB<br/>Stdev: 3.676 MB (1.0%)<br/>Runs: 351.7682291666667 353.6223958333333 354.5 354.6904296875 355.236328125 360.6940104166667 360.8059895833333 362.220703125                                                                                              |
</details>

### Fixed Issues
<!---
1. Please postfix `$` with a URL link to the GitHub issue this Pull Request is fixing. For example, `$ https://github.com/Expensify/App/issues/<issueID>`.
2. Please postfix  `PROPOSAL:` with a URL link to your GitHub comment, which contains the approved proposal (i.e. the proposal that was approved by Expensify).  For example, `PROPOSAL: https://github.com/Expensify/App/issues/<issueID>#issuecomment-1369752925`

Do NOT add the special GH keywords like `fixed` etc, we have our own process of managing the flow.
It MUST be an entire link to the github issue and your comment proposal ; otherwise, the linking and its automation will not work as expected.

Make sure this section looks similar to this (you can link multiple issues using the same formatting, just add a new line):

$ https://github.com/Expensify/App/issues/<issueID>
$ https://github.com/Expensify/App/issues/<issueID(comment)>

Do NOT only link the issue number like this: $ #<issueID>
--->
$ N/A
PROPOSAL: https://github.com/Expensify/App/pull/43482#issuecomment-2172182535


### Tests
<!---
Add a numbered list of manual tests you performed that validates your changes work on all platforms, and that there are no regressions present.
Add any additional test steps if test steps are unique to a particular platform.
Manual test steps should be written so that your reviewer can repeat and verify one or more expected outcomes in the development environment.

For example:
1. Click on the text input to bring it into focus
2. Upload an image via copy paste
3. Verify a modal appears displaying a preview of that image
--->

- [ ] run e2e pipeline and be sure it's green;

### Offline tests
<!---
Add any relevant steps that validate your changes work as expected in a variety of network states e.g. "offline", "spotty connection", "slow internet", etc. Manual test steps should be written so that your reviewer and QA testers can repeat and verify one or more expected outcomes. If you are unsure how the behavior should work ask for advice in the `#expensify-open-source` Slack channel.
--->

N/A

### QA Steps
<!---
Add a numbered list of manual tests that can be performed by our QA engineers on the staging environment to validate that your changes work on all platforms, and that there are no regressions present.
Add any additional QA steps if test steps are unique to a particular platform.
Manual test steps should be written so that the QA engineer can repeat and verify one or more expected outcomes in the staging environment.

For example:
1. Click on the text input to bring it into focus
2. Upload an image via copy paste
3. Verify a modal appears displaying a preview of that image
--->

N/A

### PR Author Checklist
<!--
This is a checklist for PR authors. Please make sure to complete all tasks and check them off once you do, or else your PR will not be merged!
-->

- [x] I linked the correct issue in the `### Fixed Issues` section above
- [x] I wrote clear testing steps that cover the changes made in this PR
    - [x] I added steps for local testing in the `Tests` section
    - [x] I added steps for the expected offline behavior in the `Offline steps` section
    - [x] I added steps for Staging and/or Production testing in the `QA steps` section
    - [x] I added steps to cover failure scenarios (i.e. verify an input displays the correct error message if the entered data is not correct)
    - [x] I turned off my network connection and tested it while offline to ensure it matches the expected behavior (i.e. verify the default avatar icon is displayed if app is offline)
    - [x] I tested this PR with a [High Traffic account](https://github.com/Expensify/App/blob/main/contributingGuides/CONTRIBUTING.md#high-traffic-accounts) against the staging or production API to ensure there are no regressions (e.g. long loading states that impact usability).
- [x] I included screenshots or videos for tests on [all platforms](https://github.com/Expensify/App/blob/main/contributingGuides/CONTRIBUTING.md#make-sure-you-can-test-on-all-platforms)
- [x] I ran the tests on **all platforms** & verified they passed on:
    - [x] Android: Native
    - [x] Android: mWeb Chrome
    - [x] iOS: Native
    - [x] iOS: mWeb Safari
    - [x] MacOS: Chrome / Safari
    - [x] MacOS: Desktop
- [x] I verified there are no console errors (if there's a console error not related to the PR, report it or open an issue for it to be fixed)
- [x] I followed proper code patterns (see [Reviewing the code](https://github.com/Expensify/App/blob/main/contributingGuides/PR_REVIEW_GUIDELINES.md#reviewing-the-code))
    - [x] I verified that any callback methods that were added or modified are named for what the method does and never what callback they handle (i.e. `toggleReport` and not `onIconClick`)
    - [x] I verified that the left part of a conditional rendering a React component is a boolean and NOT a string, e.g. `myBool && <MyComponent />`.
    - [x] I verified that comments were added to code that is not self explanatory
    - [x] I verified that any new or modified comments were clear, correct English, and explained "why" the code was doing something instead of only explaining "what" the code was doing.
    - [x] I verified any copy / text shown in the product is localized by adding it to `src/languages/*` files and using the [translation method](https://github.com/Expensify/App/blob/4bd99402cebdf4d7394e0d1f260879ea238197eb/src/components/withLocalize.js#L60)
      - [x] If any non-english text was added/modified, I verified the translation was requested/reviewed in #expensify-open-source and it was approved by an internal Expensify engineer. Link to Slack message:
    - [x] I verified all numbers, amounts, dates and phone numbers shown in the product are using the [localization methods](https://github.com/Expensify/App/blob/4bd99402cebdf4d7394e0d1f260879ea238197eb/src/components/withLocalize.js#L60-L68)
    - [x] I verified any copy / text that was added to the app is grammatically correct in English. It adheres to proper capitalization guidelines (note: only the first word of header/labels should be capitalized), and is either coming verbatim from figma or has been approved by marketing (in order to get marketing approval, ask the Bug Zero team member to add the Waiting for copy label to the issue)
    - [x] I verified proper file naming conventions were followed for any new files or renamed files. All non-platform specific files are named after what they export and are not named "index.js". All platform-specific files are named for the platform the code supports as outlined in the README.
    - [x] I verified the JSDocs style guidelines (in [`STYLE.md`](https://github.com/Expensify/App/blob/main/contributingGuides/STYLE.md#jsdocs)) were followed
- [x] If a new code pattern is added I verified it was agreed to be used by multiple Expensify engineers
- [x] I followed the guidelines as stated in the [Review Guidelines](https://github.com/Expensify/App/blob/main/contributingGuides/PR_REVIEW_GUIDELINES.md)
- [x] I tested other components that can be impacted by my changes (i.e. if the PR modifies a shared library or component like `Avatar`, I verified the components using `Avatar` are working as expected)
- [x] I verified all code is DRY (the PR doesn't include any logic written more than once, with the exception of tests)
- [x] I verified any variables that can be defined as constants (ie. in CONST.js or at the top of the file that uses the constant) are defined as such
- [x] I verified that if a function's arguments changed that all usages have also been updated correctly
- [x] If any new file was added I verified that:
    - [x] The file has a description of what it does and/or why is needed at the top of the file if the code is not self explanatory
- [x] If a new CSS style is added I verified that:
    - [x] A similar style doesn't already exist
    - [x] The style can't be created with an existing [StyleUtils](https://github.com/Expensify/App/blob/main/src/styles/utils/index.ts) function (i.e. `StyleUtils.getBackgroundAndBorderStyle(theme.componentBG)`)
- [x] If the PR modifies code that runs when editing or sending messages, I tested and verified there is no unexpected behavior for all supported markdown - URLs, single line code, code blocks, quotes, headings, bold, strikethrough, and italic.
- [x] If the PR modifies a generic component, I tested and verified that those changes do not break usages of that component in the rest of the App (i.e. if a shared library or component like `Avatar` is modified, I verified that `Avatar` is working as expected in all cases)
- [x] If the PR modifies a component related to any of the existing Storybook stories, I tested and verified all stories for that component are still working as expected.
- [x] If the PR modifies a component or page that can be accessed by a direct deeplink, I verified that the code functions as expected when the deeplink is used - from a logged in and logged out account.
- [x] If the PR modifies the UI (e.g. new buttons, new UI components, changing the padding/spacing/sizing, moving components, etc) or modifies the form input styles:
    - [x] I verified that all the inputs inside a form are aligned with each other.
    - [x] I added `Design` label and/or tagged `@Expensify/design` so the design team can review the changes.
- [x] If a new page is added, I verified it's using the `ScrollView` component to make it scrollable when more elements are added to the page.
- [x] If the `main` branch was merged into this PR after a review, I tested again and verified the outcome was still expected according to the `Test` steps.

### Screenshots/Videos
<details>
<summary>Android: Native</summary>

<!-- add screenshots or videos here -->

</details>

<details>
<summary>Android: mWeb Chrome</summary>

<!-- add screenshots or videos here -->

</details>

<details>
<summary>iOS: Native</summary>

<!-- add screenshots or videos here -->

</details>

<details>
<summary>iOS: mWeb Safari</summary>

<!-- add screenshots or videos here -->

</details>

<details>
<summary>MacOS: Chrome / Safari</summary>

<!-- add screenshots or videos here -->

</details>

<details>
<summary>MacOS: Desktop</summary>

<!-- add screenshots or videos here -->

</details>