@hayata-suenaga 2024-06-28
https://github.com/Expensify/App/pull/43482 was merged, so this one is ready to be merged 😄