@OSBotify 2024-07-15
🚀 [Deployed](https://github.com/Expensify/App/actions/runs/9945901401) to production by https://github.com/thienlnam in version: 9.0.6-8 🚀

platform | result
---|---
🤖 android 🤖|success ✅
🖥 desktop 🖥|success ✅
🍎 iOS 🍎|success ✅
🕸 web 🕸|success ✅